consolidate exceptions
