+++ corehq/apps/case_search/xpath_functions/__init__.py
@@ -1,6 +1,6 @@
 from .query_functions import _not
 from .subcase_functions import subcase
-from .value_functions import XPathFunctionException, date  # noqa: F401
+from .value_functions import date  # noqa: F401
 
 # functions that transform a value
 XPATH_VALUE_FUNCTIONS = {
