--- corehq/apps/case_search/xpath_functions/exceptions.py
@@ -1,2 +1,0 @@
-class XPathFunctionException(Exception):
-    pass
